refactor: Prep for SS-10
