+++ src/components/app/AppLayout.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import { useRouter } from 'next/router';
-import { SidebarProvider } from '@/components/ui/sidebar';
-import { AppSidebar } from '@/components/AppSidebar';
-import Header from '@/components/Header';
+import { SidebarProvider } from '@/components/app/ui/sidebar';
+import { AppSidebar } from '@/components/app/AppSidebar';
+import AppHeader from '@/components/app/AppHeader';
 
 
 type LayoutProps = {
     children: React.ReactNode;
 };
 
-const Layout: React.FC<LayoutProps> = ({ children }) => {
+const AppLayout: React.FC<LayoutProps> = ({ children }) => {
     const router = useRouter();
     const isHomepage = router.pathname === '/';
 
@@ -17,11 +17,11 @@ const Layout: React.FC<LayoutProps> = ({ children }) => {
         <SidebarProvider>
             {!isHomepage && <AppSidebar />}
             <main className="w-screen h-screen bg-white">
-                {!isHomepage && <Header />}
+                {!isHomepage && <AppHeader />}
                 {children}
             </main>
         </SidebarProvider>
     );
 };
 
-export default Layout;
+export default AppLayout;
