+++ src/components/app/ui/fade-text.tsx
